# docs/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "KnxMonitor.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Knx/status": {
      "get": {
        "tags": [
          "Knx"
        ],
        "summary": "Current state of the KNX bus link.",
        "description": "Read-only — nothing is sent to the bus. Configuration is the configuration the live link was established with and is null while disconnected; it is not necessarily the stored configuration flagged active. ManualDisconnect is latched by a manual disconnect and keeps the auto-connect worker from reconnecting until the next connect clears it.",
        "responses": {
          "200": {
            "description": "Connection flag, link state (Disconnected / Connecting / Connected), the manual-disconnect latch and the configuration currently in use."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Knx/connect": {
      "post": {
        "tags": [
          "Knx"
        ],
        "summary": "Opens the KNX bus link using the stored configuration with the given id.",
        "description": "Touches the installation: an IP tunnel is opened (or the routing multicast group joined), and received telegrams start flowing into the live view and the history again. An existing link is torn down first. Clears the manual-disconnect latch, so the auto-connect worker may keep the link up from here on. The request body is the bare configuration id (a JSON number). Unknown id yields 404, a failed connect 400.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "text/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Knx/disconnect": {
      "post": {
        "tags": [
          "Knx"
        ],
        "summary": "Closes the KNX bus link.",
        "description": "Ends telegram recording until the next connect. Latches manual-disconnect so the auto-connect worker does not reconnect on its own; the next connect (or enabling auto-connect) clears the latch. Succeeds even when no link is currently up.",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Knx/test-connection": {
      "post": {
        "tags": [
          "Knx"
        ],
        "summary": "Probes whether a stored configuration can reach the bus. Sends no group telegram and changes nothing in the stored configuration.",
        "description": "The probe uses a throwaway connection. If a live link is currently up it is paused for the duration of the probe (freeing the gateway's tunnel slot) and restored afterwards, so recording has a short gap — the frontend warns before triggering this while live. For routing configurations the probe additionally listens up to five seconds for real bus traffic: joining a multicast group succeeds locally even when no telegram can ever arrive. That case is reported as JoinedWithoutTraffic and counts as a warning, not a failure (Success stays true). Unknown id yields 404.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "text/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success flag (false only for Failed), the outcome and a human-readable message."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Knx/autoconnect": {
      "get": {
        "tags": [
          "Knx"
        ],
        "summary": "Auto-connect state for the effective configuration (active one, else the first). The active project is coupled to the bus link, so this governs whether it auto-connects.",
        "description": "Read-only. When no configuration is stored at all, Enabled is reported as true (the default for a configuration that has yet to be created).",
        "responses": {
          "200": {
            "description": "The auto-connect flag and whether a bus link is currently up."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "Knx"
        ],
        "summary": "Enables or disables auto-connect for the effective configuration.",
        "description": "Applied immediately, not merely stored: enabling opens the bus link right away when none is up (which clears the manual-disconnect latch), disabling closes an existing link. A failed immediate connect is not reported as an error — the auto-connect worker retries on its next tick (roughly every ten seconds). 404 when no configuration is stored.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAutoConnectRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAutoConnectRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetAutoConnectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The stored flag and whether a bus link is up afterwards."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Knx/write": {
      "post": {
        "tags": [
          "Knx"
        ],
        "summary": "Sends a GroupValueWrite to the bus. ⚠ Writes to the real KNX installation — the frontend gates this behind an explicit confirmation dialog.",
        "description": "This has a physical effect: the addressed actuators act on the value. The value is encoded with the given datapoint type; without a DatapointType only a raw hex string (\"0x01\", \"01 02\") is accepted, anything else is rejected. Answers 400 with Success=false when the address is missing or unparsable, when no bus link is up, or when the value could not be encoded for the given DPT.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WriteGroupValueRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WriteGroupValueRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WriteGroupValueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Knx/read": {
      "post": {
        "tags": [
          "Knx"
        ],
        "summary": "Sends a GroupValueRead; the response shows up as a normal telegram.",
        "description": "Puts a telegram on the real installation, but changes no value — the addressed device answers with its current one. The answering GroupValueResponse arrives through the normal receive path and appears in the live view and the history; this endpoint itself only confirms that the request was sent, it does not return the value. Answers 400 with Success=false when the address is missing or unparsable, or when no bus link is up.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadGroupValueRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadGroupValueRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReadGroupValueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Knx/configurations": {
      "post": {
        "tags": [
          "Knx"
        ],
        "summary": "Stores a new KNX interface configuration.",
        "description": "Storing alone does not touch the bus: the entry is created inactive and no connection is opened here. Connect explicitly, or leave AutoConnect on — the auto-connect worker picks a configuration only while no link is up, preferring the active one, else the first with AutoConnect enabled. The payload carries no secure-tunnel flag, so that option stays off for new entries.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConfigRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConfigRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The stored configuration including its assigned id."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "get": {
        "tags": [
          "Knx"
        ],
        "summary": "Lists all stored KNX interface configurations.",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Knx/configurations/{id}": {
      "get": {
        "tags": [
          "Knx"
        ],
        "summary": "Returns a single stored KNX interface configuration.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the stored configuration; 404 when it does not exist.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "Knx"
        ],
        "summary": "Updates a stored KNX interface configuration; may re-establish the live bus link.",
        "description": "When the gateway endpoint actually changes (IP address, port or connection type), the configuration has AutoConnect enabled and a link is currently up, the running connection is re-opened with the new settings — that briefly interrupts recording. Changing only the physical address or the auto-connect flag does not trigger a reconnect. A failed reconnect is not an error: the response reports Reconnected=false and the auto-connect worker retries. The payload carries no secure-tunnel flag, so that setting stays as stored.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the stored configuration; 404 when it does not exist.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConfigRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConfigRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The stored configuration and whether the live link was re-established."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ConnectionType": {
        "type": "integer"
      },
      "CreateConfigRequest": {
        "required": [
          "ipAddress",
          "port",
          "physicalAddress",
          "connectionType"
        ],
        "type": "object",
        "properties": {
          "ipAddress": {
            "type": "string"
          },
          "port": {
            "type": "integer",
            "format": "int32"
          },
          "physicalAddress": {
            "type": "string"
          },
          "connectionType": {
            "$ref": "#/components/schemas/ConnectionType"
          },
          "autoConnect": {
            "type": "boolean",
            "default": true
          }
        }
      },
...
      "ReadGroupValueRequest": {
        "required": [
          "address"
        ],
        "type": "object",
        "properties": {
          "address": {
            "type": "string"
          }
        }
      },
...
      "SetAutoConnectRequest": {
        "required": [
          "enabled"
        ],
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          }
        }
      },
      "WriteGroupValueRequest": {
        "required": [
          "address",
          "datapointType",
          "value"
        ],
        "type": "object",
        "properties": {
          "address": {
            "type": "string"
          },
          "datapointType": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Access-Token aus POST /api/Auth/login (ohne \"Bearer \"-Präfix einfügen).",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}